Rationalized data_external functions.
