+++ .githooks/data_external/generate_db_connection_string.py
@@ -3,7 +3,6 @@ import os
 import json
 import sys
 from types import SimpleNamespace
-from typing import List
 
 sys.dont_write_bytecode = True
 
@@ -19,18 +18,18 @@ sys.dont_write_bytecode = True
 project_root = os.getcwd()
 os.chdir(f"{project_root}/.githooks")
 sys.path.append(".")
-from utils.extractors import get_tfvars_as_json
-#from utils.logger import logger
-from utils.logger import external_logger #as logger
 
-# Extract tfvars just like we do with the Python validation script
+from utils.extractors import get_tfvars_as_json #convert_tfvars_to_dictionary
+from utils.logger import external_logger
+from utils.common_data_external_functions import getDVal, return_tf_payload
+
 os.chdir(project_root)
 data_dictionary = get_tfvars_as_json()
 data = SimpleNamespace(**data_dictionary)
 
 # Much simpler way to get variable passed in (via Terraform sending to stdin)
 query = json.load(sys.stdin)
-query = SimpleNamespace(**query)
+# query = SimpleNamespace(**query)
 ## ------------------------------------------------------------------------------------
 
 
@@ -42,27 +41,11 @@ mysql8_connstring = "allowPublicKeyRetrieval=true&useSSL=false"
 v24plus_connstring = "permitMysqlScheme=true"
 
 
-def return_tf_payload(status: str, value: str):
-    external_logger.debug(f"Value is: {value}")
-    payload = {'status': status, 'value': value}
-    external_logger.debug(f"Payload is: {payload}")
-
-    print(json.dumps(payload))
-    # MemoryHandler configured to flush on `logger.error`. Flush once after TF gets its payload (above).
-    external_logger.error("Flushing.")     #external_logger.flush()
-    exit(0)
-
-
 def generate_connection_string(mysql8: str, v24plus: str):
     connection_string = ""
-    add_mysql8 = False
-    add_v24plus = False
 
-    if mysql8.startswith("8."): 
-        add_mysql8 = True
-
-    if v24plus >= "v24":
-        add_v24plus = True
+    add_mysql8 = True if mysql8.startswith("8.") else False
+    add_v24plus = True if v24plus >= "v24" else False
 
     if add_mysql8 and add_v24plus:
         connection_string = f"?{mysql8_connstring}&{v24plus_connstring}"
@@ -71,9 +54,13 @@ def generate_connection_string(mysql8: str, v24plus: str):
     elif not add_mysql8 and add_v24plus:
         connection_string = f"?{v24plus_connstring}"
 
-    return connection_string
+    values = {
+        "connection_string": connection_string
+    }
+
+    return values
 
 
 if __name__ == '__main__':
-    connection_string = generate_connection_string(engine_version, data.tower_container_version)
-    return_tf_payload("0", connection_string)
+    values = generate_connection_string(engine_version, data.tower_container_version)
+    return_tf_payload("0", values)
