+++ .githooks/data_external/generate_dns_values.py
@@ -3,7 +3,6 @@ import os
 import json
 import sys
 from types import SimpleNamespace
-from typing import Any, List
 
 sys.dont_write_bytecode = True
 
@@ -19,10 +18,11 @@ sys.dont_write_bytecode = True
 project_root = os.getcwd()
 os.chdir(f"{project_root}/.githooks")
 sys.path.append(".")
+
 from utils.extractors import get_tfvars_as_json #convert_tfvars_to_dictionary
 from utils.logger import external_logger
+from utils.common_data_external_functions import getDVal, return_tf_payload
 
-# Extract tfvars just like we do with the Python validation script
 os.chdir(project_root)
 data_dictionary = get_tfvars_as_json()
 data = SimpleNamespace(**data_dictionary)
@@ -32,13 +32,6 @@ query = json.load(sys.stdin)
 # query = SimpleNamespace(**query)
 ## ------------------------------------------------------------------------------------
 
-
-# https://www.reddit.com/r/learnpython/comments/y02net/is_there_a_better_way_to_store_full_dictionary/
-def getDVal(d : dict, listPath : list) -> Any:
-    '''Recursively loop through a dictionary object to get to the target nested key.'''
-    for key in listPath:
-        d = d[key]
-    return d
 
 # Vars
 dns_zone_mappings = {
@@ -65,10 +58,6 @@ def populate_values(query):
     dns_zone_id = ""
     dns_instance_ip = ""
 
-    # Determine kinda of DNS record to create
-    # dns_create_alb_record = True if (data.flag_create_load_balancer and not data.flag_create_hosts_file_entry) else False
-    # dns_create_ec2_record = True if (not data.flag_create_load_balancer and not data.flag_create_hosts_file_entry) else False
-
     for k,v in dns_zone_mappings.items():
         external_logger.debug(f"k is : {k}; and v is: {v}")
         tf_obj, dpath = v
@@ -83,33 +72,11 @@ def populate_values(query):
         dns_instance_ip = getDVal(tf_obj_json, dpath) if data_dictionary[k] else dns_instance_ip
 
     values = {
-        # "dns_create_alb_record": dns_create_alb_record,
-        # "dns_create_ec2_record": dns_create_ec2_record,
-
         "dns_zone_id": dns_zone_id,
         "dns_instance_ip": dns_instance_ip
     }
 
     return values
-
-
-def convert_booleans_to_strings(payload: dict) -> dict: 
-    for k,v in payload.items():
-        external_logger.debug(f"k is {k} and v is {v}.")
-        if isinstance(v, bool):
-            payload[k] = "true" if v == True else "false"
-    return payload
-        
-
-def return_tf_payload(status: str, values: dict):
-    external_logger.debug(f"Payload is: {values}")
-
-    payload = {'status': status, **values}
-    payload = convert_booleans_to_strings(payload)
-    print(json.dumps(payload))
-
-    external_logger.error("Flushing.")
-    exit(0)
 
 
 if __name__ == '__main__':
