+++ .githooks/data_external/generate_flags.py
@@ -3,7 +3,6 @@ import os
 import json
 import sys
 from types import SimpleNamespace
-from typing import Any, List
 
 sys.dont_write_bytecode = True
 
@@ -19,10 +18,11 @@ sys.dont_write_bytecode = True
 project_root = os.getcwd()
 os.chdir(f"{project_root}/.githooks")
 sys.path.append(".")
+
 from utils.extractors import get_tfvars_as_json #convert_tfvars_to_dictionary
 from utils.logger import external_logger
+from utils.common_data_external_functions import getDVal, return_tf_payload
 
-# Extract tfvars just like we do with the Python validation script
 os.chdir(project_root)
 data_dictionary = get_tfvars_as_json()
 data = SimpleNamespace(**data_dictionary)
@@ -31,7 +31,6 @@ data = SimpleNamespace(**data_dictionary)
 query = json.load(sys.stdin)
 # query = SimpleNamespace(**query)
 ## ------------------------------------------------------------------------------------
-
 
 
 def populate_values(query):
@@ -48,25 +47,6 @@ def populate_values(query):
     }
 
     return values
-
-
-def convert_booleans_to_strings(payload: dict) -> dict: 
-    for k,v in payload.items():
-        external_logger.debug(f"k is {k} and v is {v}.")
-        if isinstance(v, bool):
-            payload[k] = "true" if v == True else "false"
-    return payload
-        
-
-def return_tf_payload(status: str, values: dict):
-    external_logger.debug(f"Payload is: {values}")
-
-    payload = {'status': status, **values}
-    payload = convert_booleans_to_strings(payload)
-    print(json.dumps(payload))
-
-    external_logger.error("Flushing.")
-    exit(0)
 
 
 if __name__ == '__main__':
