+++ 000_main.tf
@@ -64,12 +64,7 @@ data "aws_subnet" "existing" {
 # https://medium.com/@leslie.alldridge/terraform-external-data-source-using-custom-python-script-with-example-cea5e618d83e
 data "external" "generate_db_connection_string" {
   program = ["python3", "${path.module}/.githooks/data_external/generate_db_connection_string.py"]
-  #query = {
-    # tower_container_version = var.tower_container_version
-    # flag_use_container_db = var.flag_use_container_db
-    # db_container_engine_version = var.db_container_engine_version
-    # db_engine_version = var.db_engine_version
-  #}
+  query = {}
 }
 
 data "external" "generate_flags" {
@@ -241,7 +236,7 @@ locals {
 
   # tower_db_url = var.flag_create_external_db == true ? module.rds[0].db_instance_address : var.tower_db_url
   tower_db_root = ( var.flag_use_container_db == true? var.tower_db_url : module.rds[0].db_instance_address )
-  tower_db_url = "${local.tower_db_root}/${var.db_database_name}${data.external.generate_db_connection_string.result.value}"
+  tower_db_url = "${local.tower_db_root}/${var.db_database_name}${data.external.generate_db_connection_string.result.connection_string}"
 
 
   # Redis
