+++ 012_outputs.tf
@@ -27,5 +27,5 @@ output "redis_endpoint" {
 # Example for how to get values dynamically generated by `data.external`
 output "database_connection_string" {
   description = "Dynamically generated db connectino string based on tfvars selections."
-  value = data.external.generate_db_connection_string.result.value
+  value = data.external.generate_db_connection_string.result.connection_string
 }
